Plug up all non-void functions without return statements
Only a void function is allowed to have a return statement: otherwise, the warning can prove to be fatal post-compilation as the routine runs past the end of the code into some other undefined territory.

This commit should resolve every single instance of the warning "control may reach end of non-void function" (whether it always/sometimes does so or not).